Expose Channel publicly
