+++ module/camera/protected/kipr/camera/channel_impl.hpp
@@ -1,8 +1,9 @@
-#ifndef _KIPR_CAMERA_CHANNEL_HPP_
-#define _KIPR_CAMERA_CHANNEL_HPP_
+#ifndef _KIPR_CAMERA_CHANNEL_IMPL_HPP_
+#define _KIPR_CAMERA_CHANNEL_IMPL_HPP_
 
 #include "kipr/camera/object.hpp"
 #include "kipr/config/config.hpp"
+#include "kipr/camera/channel.hpp"
 
 #include <opencv2/core/core.hpp>
 
@@ -39,29 +40,6 @@ namespace kipr
     private:
       // TODO: private constructor?
       static std::map<std::string, ChannelImpl *> m_channelImpls;
-    };
-
-    class Channel
-    {
-    public:
-      Channel(Device *device, const config::Config &config);
-      ~Channel();
-
-      void invalidate();
-      const ObjectVector *objects() const;
-      Device *device() const;
-
-      /**
-      * Do not call this method unless you know what you are doing!
-      */
-      void setConfig(const config::Config &config);
-
-    private:
-      Device *m_device;
-      config::Config m_config;
-      mutable ObjectVector m_objects;
-      ChannelImpl *m_impl;
-      mutable bool m_valid;
     };
 
     typedef std::vector<Channel *> ChannelPtrVector;
